+++ src/store/combineReducers.js
@@ -1,6 +1,6 @@
 import { combineReducers } from "redux";
 import users from "./users";
-import error from "./error";
+import error from "./login/error";
 import loggedInUser from "./loggedInUser";
 
 export default combineReducers({
