+++ src/store/login/users.js
@@ -4,8 +4,6 @@ export default function(state = [], action = {}) {
   switch (action.type) {
     case ALL_USERS:
       return action.payload;
-    case NEW_USER:
-      return [...state, action.payload];
     default:
       return state;
   }
